python/requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
attrs==25.4.0
beautifulsoup4==4.12.3
blinker==1.9.0
ccxt==4.4.6
certifi==2024.6.2
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.3.0
contourpy==1.3.3
cryptography==46.0.3
curl_cffi==0.7.1
cycler==0.12.1
dateparser==1.2.0
defusedxml==0.7.1
distro==1.9.0
Flask==3.0.3
fonttools==4.60.1
fpdf2==2.7.9
frozendict==2.4.4
frozenlist==1.8.0
git-filter-repo==2.47.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
kaleido==0.2.1
kiwisolver==1.4.9
lxml==6.0.2
Markdown==3.6
MarkupSafe==3.0.3
matplotlib==3.9.2
multidict==6.7.0
multitasking==0.0.11
numpy==1.26.4
openai==2.7.1
packaging==25.0
pandas==2.2.2
patsy==1.0.2
peewee==3.17.5
pillow==12.0.0
platformdirs==4.2.2
plotly==5.24.1
propcache==0.4.1
protobuf==5.27.1
pycares==4.11.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.10
pydantic_core==2.33.2
pyparsing==3.2.5
python-binance==1.0.29
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2025.11.3
reportlab==4.2.2
requests==2.32.3
scikit-learn==1.4.2
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
statsmodels==0.14.5
tenacity==9.1.2
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.3.1
urllib3==2.2.2
webencodings==0.5.1
websockets==10.4
Werkzeug==3.1.3
yarl==1.22.0
yfinance==0.2.40
